refactor(bayn): make simulation reconciliation functional
